Report stdout if not empty after khiops execution
The Khiops executable normally don't emit messages to stdout since we
use the log files. However there can be messages in stdout coming either
from MPI or the filesystems plugins.

So when there is a non empty stdout we report it as a warning if the
excution went ok or in the runtime error if not.