Fix Python 3.8 test failures
In Python 3.8, `scipy.sparse.lil_matrix` rows contain arrays of Python lists.
If empty row, it contains a single-element array with an empty Python
list element; its numpy array data has size 1.
This patch tests directly on the size of the row and flattens the rows
to ensure a homogeneous treatment.