Set KhiopsEncoder to "unfit" when there are no informative vars
If there are no informative vars for `KhiopsEncoder` we declare the fit
as a failed an undefine all attributes.

Note that we use the sklearn convention that an estimator is fitted if
and only if no attributes are defined. Estimator attributes are
conventionally defined as public attributes whose name have a suffix
`_`.