Control the types on sklearn internal read table
We do this only for KhiopsClassifier and KhiopsRegressor: It is critical
for KhiopsClassifier as it accepts many target types and it is trivial
in the case of KhiopsRegressor.

For KhiopsEncoder and KhiopsCoclustering is less critical and for the
first one it is very complex. We left them as TODO's.

Additionaly, we now also check in the "output type" tests that the
result of predict is correct. Before we only checked only that the
classes_ attribute was ok. This is to further ensure correctness.