Skip baseline model, always present in regression, absent from the others
The regression model .kdic now always includes the baseline model. The
other estimators' models now never include the baseline model.

For this "MVP" beta, we keep track of the baseline model if present, but
we currently ignore it, even for regression; thus, as before, only the
SNB model is used in the estimators.